requirements.txt
paho-mqtt==2.1.0
boto3==1.38.8
debugpy==1.8.14
duckdb==1.2.2
numpy==1.26.4
pandas==2.2.2
streamlit
psycopg2
bcrypt
faker
minio